refactors form structure
