+++ src/views/Login.vue
@@ -4,7 +4,7 @@
   </div>
 </template>
 <script>
-import LoginForm from "@/components/LoginForm.vue";
+import LoginForm from "@/components/forms/LoginForm.vue";
 export default {
   name: "App",
   components: {
