+++ src/views/SignUp.vue
@@ -4,7 +4,7 @@
   </div>
 </template>
 <script>
-import SignUpForm from "@/components/SignUpForm.vue";
+import SignUpForm from "@/components/forms/SignUpForm.vue";
 export default {
   name: "App",
   components: {
